requirements.txt
boto3==1.20.24
canvasapi==2.2.0
cryptography==37.0.2
decorator==5.1.1
eventlet==0.30.2
Flask-Login==0.6.1
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask==2.1.2
gevent==21.12.0
gevent-websocket==0.10.1
gunicorn==20.1.0
lti==0.9.5
oauthlib==3.1.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-magic==0.4.26
pytz==2022.1
requests==2.28.1
simplejson==3.17.6
smart-open==4.2.0
SQLAlchemy==1.4.39
sqlvalidator==0.0.19
Werkzeug==2.1.2
https://github.com/python-cas/python-cas/archive/master.zip

# For testing

moto==3.1.16
pytest==7.1.2
pytest-flask==1.2.0
responses==0.21.0
tox==3.25.1
